moves api  methods to src root
