+++ web/src/routes/settlementSelector/selector.tsx
@@ -1,5 +1,5 @@
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { Get } from "@/api/api";
+import { Get } from "@/api";
 import { Settlement } from "@/types";
 import { CreateSettlementDialogue } from "./createSettlementDialogue";
 import { Button } from "@/components/ui/button";
